Paying for College - JS Reorganization - Move JS files into subfolder - Add string protections
