+++ cfgov/unprocessed/apps/paying-for-college/js/college-costs/dispatchers/update-models.js
@@ -187,6 +187,7 @@ function updateSchoolData( iped ) {
       } )
       .catch( function( error ) {
         reject( error );
+        iped = iped.replace(/\D/g,'');
         console.log( 'An error occurred when accessing school data for ' + iped, error );
       } );
   } );
