+++ cfgov/unprocessed/apps/paying-for-college/js/college-costs/util/analytics.js
@@ -1,4 +1,4 @@
-import Analytics from '../../../../js/modules/Analytics';
+import Analytics from '../../../../../js/modules/Analytics';
 
 /**
  * Sends an event to the dataLayer for Google Tag Manager
